Removed various integer overflows for Wirehair
There are two things from clang that I don't quite understand how they
can happen, so I want to solve those and experiment a bit before
releasing v0.2.3

+++ src/bef.h
@@ -206,7 +206,7 @@ void bef_encode_free(char **data, char **parity, uint16_t k, uint16_t m);
  *
  * error codes not yet defined, but will return 0 when successful
  */
-int bef_decode_ecc(char **frags, uint16_t frag_len, size_t frag_b,
+int bef_decode_ecc(char **frags, uint32_t frag_len, size_t frag_b,
 		   char **output, size_t *onbyte,
 		   struct bef_real_header header);
